fix(test): .pubsub.peers
