+++ test/interface-tests/src/pubsub/utils.js
@@ -1,5 +1,6 @@
 import { nanoid } from 'nanoid'
 import delay from 'delay'
+import pRetry from 'p-retry'
 
 /**
  * @param {import('ipfs-core-types').IPFS} ipfs
@@ -25,6 +26,36 @@ export async function waitForPeers (ipfs, topic, peersToWait, waitForMs) {
 
     await delay(10)
   }
+}
+
+const retryOptions = {
+  retries: 5,
+  onFailedAttempt: async ({ attemptNumber }) => {
+    await delay(1000 * attemptNumber)
+  },
+  maxRetryTime: 10000
+}
+
+/**
+ * @param {string} topic
+ * @param {import('ipfsd-ctl').Controller["peer"]} peer
+ * @param {import('ipfsd-ctl').Controller} daemon
+ * @param {Parameters<typeof pRetry>[1]} rOpts
+ */
+export const waitForTopicPeer = (topic, peer, daemon, rOpts = {}) => {
+  return pRetry(async () => {
+    // console.log(`waiting for topic ${topic} from peer ${peer.id.toString()} on ${daemon.peer.id.toString()}`)
+    const peers = await daemon.api.pubsub.peers(topic)
+
+    if (!peers.map(p => p.toString()).includes(peer.id.toString())) {
+      throw new Error(`Could not find peer ${peer.id}`)
+    } else {
+      // console.log(`Peer found for topic ${topic}`)
+    }
+  }, {
+    retryOptions,
+    ...rOpts
+  })
 }
 
 export function getTopic () {
